Prevent warnings when English translations can not be load
English translation files are built-in and usually only used
for plural forms. Hence the files are often empty and the
warnings about them are more misleading than helpful most of
the time.

TODO: Find a way to distinguish between translations
which are supposed to be empty and missing/invalid
translations.